+++ NovaSass.novaextension/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## Version 0.9
+- Auto hide the error notification after 10 seconds or if the file is re-saved without an error.
+- Other bug fixes
+- See README Notes for a change to the compile directory
+- Updated README
+
+## Version 0.8
+- Changed the update root from the workspace root to the files current directory
+
+## Version 0.7
+- Updated Icon
+
 ## Version 0.6.5
 - Bug fix if only one line of message